refactor(core): extract action-engine type surface + helpers into sibling modules (#462)
Split action-engine.ts (2046 -> 1679 lines) by responsibility:
- Public contract types (DataProvider, ActionExecutor, ExecuteOptions,
  ActionExecutorOptions, ...) -> engine/action-engine-types.ts
- Module-private helpers (isExecutionMeta, fillMissingSystemKeys,
  checkAndRunFieldLockInterceptor) -> engine/action-helpers.ts

action-engine.ts re-exports every type for backward compatibility, so all
existing imports from @linchkit/core / ./action-engine keep working
unchanged. Pure code movement, zero behavior change (typecheck + full
test suite green).
